+++ src/utils/Utils.ts
@@ -1,4 +1,4 @@
-import { writeFileSync } from 'node:fs';
+import { writeFileSync } from 'fs';
 
 import Fetcher from '@/core/Fetcher';